Create Locker Batch File
This Locker Batch File is a simple script that allows users to create a password-protected folder in Windows. It works by creating a "Private" folder where users can store sensitive files. When the script is executed, users can choose to lock the folder, making it hidden and inaccessible. To unlock it, the user must enter the correct password.