Create To-Do List Manager Batch File
The Simple To-Do List Manager is a batch script that allows users to manage tasks in a text file. Users can add tasks, view tasks, and delete all tasks through a simple menu-driven interface. The tasks are stored in tasks.txt, making it easy to keep track of to-do items.