Create Fake Error Message Batch File
This Batch File script displays a **custom error message** in a pop-up window to prank users.  
It uses the `msg` command to send a message to the screen, simulating a fake system warning or error.  
This is a harmless and fun way to **fool your friends or coworkers** for a laugh